requirements.txt
# === Core LangChain ecosystem (early 2025 compatible) ===
langchain==0.3.16
langchain-core>=0.3.32,<0.4.0
langchain-community>=0.3.10,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-chroma>=0.1.4,<0.2.0
langchain-text-splitters>=0.3.0,<0.4.0
chromadb==0.5.10

# === ML / CV stack ===
torch==2.2.2
torchvision
transformers==4.39.3
scikit-learn==1.4.2
numpy==1.26.4
Pillow==10.2.0
tqdm==4.66.4
plotly==5.20.0

# === Utilities ===
PyPDF2==3.0.1
python-box==7.2.0
python-dotenv
ensure
gradio

#### USE PYTHON 3.11 ####
#### DO NOT CHANGE ANY VERSION ####